+++ swarm/src/dummy.rs
@@ -108,7 +108,8 @@ impl crate::handler::ConnectionHandler for ConnectionHandler {
             | ConnectionEvent::ListenUpgradeError(_)
             | ConnectionEvent::LocalProtocolsChange(_)
             | ConnectionEvent::RemoteProtocolsChange(_)
-            | ConnectionEvent::Datagram(_) => {}
+            | ConnectionEvent::Datagram(_)
+            | ConnectionEvent::DatagramMaxSize(_) => {}
         }
     }
 }