+++ swarm/src/handler/one_shot.rs
@@ -190,7 +190,8 @@ where
             | ConnectionEvent::ListenUpgradeError(_)
             | ConnectionEvent::LocalProtocolsChange(_)
             | ConnectionEvent::RemoteProtocolsChange(_)
-            | ConnectionEvent::Datagram(_) => {}
+            | ConnectionEvent::Datagram(_)
+            | ConnectionEvent::DatagramMaxSize(_) => {}
         }
     }
 }